Refactoring user & auth
