+++ src/services/user.service.ts
@@ -1,34 +1,81 @@
 import {User, PlatformType} from '../entities/user.entity';
 
-export async function getUsers() {
-  const users: User[] = await User.find({withDeleted: false});
-  return users;
-}
+// export async function getUsers() {
+//   const users: User[] = await User.find({withDeleted: false});
+//   return users;
+// }
 
 export async function getUser(id: number) {
-  const user: User = await User.findOne({id});
+  const user: User = await User.findOne(
+    {id},
+    {
+      select: [
+        'id',
+        'sub',
+        'email',
+        'nickname',
+        'platformType',
+        'deletedAt',
+        'createdAt',
+        'updatedAt',
+      ],
+    }
+  );
   return user;
 }
 
-export async function addUser(nickname: string, platformType?: PlatformType) {
+export async function addUser(
+  sub: string,
+  email: string,
+  nickname: string,
+  platformType: PlatformType
+) {
   const user: User = new User();
+  user.sub = sub;
   user.nickname = nickname;
-  if (platformType) {
-    user.platformType = platformType;
-  }
+  user.email = email;
+  user.platformType = platformType;
   await user.save();
   return user;
 }
 
 export async function updateUser(id: number, nickname: string) {
-  const user: User = await User.findOne({id});
+  const user: User = await User.findOne(
+    {id},
+    {
+      select: [
+        'id',
+        'sub',
+        'email',
+        'nickname',
+        'platformType',
+        'deletedAt',
+        'createdAt',
+        'updatedAt',
+      ],
+    }
+  );
   user.nickname = nickname;
   await user.save();
   return user;
 }
 
 export async function deleteUser(id: number) {
-  const user: User = await User.findOne({id});
+  const user: User = await User.findOne(
+    {id},
+    {
+      select: [
+        'id',
+        'sub',
+        'email',
+        'nickname',
+        'platformType',
+        'deletedAt',
+        'createdAt',
+        'updatedAt',
+      ],
+    }
+  );
   await user.softRemove();
   return user;
 }
